Fix false event diffs Events are tricky to handle right, for two reasons:
1) STARTS and ENDS can be expressions in your SQL but are stored in MySQL as
absolute dates, computed at the insert time.

2) We instantiate a status of DISABLE ON SLAVE such that when we're done
it's set to ENABLED in MySQL. This is a weirdo local DBA custom that can
probably be done away with.

This fixes both of these, by storing STARTS and ENDS as metadata, since
MySQL won't round trip them for us, and treating DISABLE ON SLAVE in the DB
you're migrating to as equivalent to ENABLED in the one you're coming from.